Image Resize by ID
Now resizing an image based on an attachment ID, not based on an url. 
In Matthews original version, it needed to make a DB query checking for if the GUID exists in the WPdb. 
Since we're using Media File renamer, this gave some problems. So now based on the ID, we can do with it whatever we want. 
Added a function to get the right proportions if crop turned off and no dimensions where given.  
Removed WP support below 3.5, since we're using the latest version. F** those old ones. 
Added support for WPML translated attachments. Not really necessary probably, since we're just returning the url. (Which is always the same). 
But heey. Atleast it will work great. 

Also changed the return to an array or a string just containing the url, for easy implementation.